RNG: work with more than int32_t
will hopefully help in silencing some of the conversion warnings

+++ gemrb/plugins/OpenALAudio/AmbientMgrAL.h
@@ -61,7 +61,7 @@ private:
 		const Ambient* ambient;
 		uint64_t lastticks;
 		unsigned int nextdelay;
-		unsigned int nextref;
+		size_t nextref;
 		unsigned int totalgain;
 
 		bool isHeard(const Point &listener) const;
